refactor: resolver 패키지 이동
